+++ classes/editor.class.php
@@ -2,7 +2,7 @@
 // +-------------------------------------------------+
 // © 2002-2004 PMB Services / www.sigb.net pmb@sigb.net et contributeurs (voir www.sigb.net)
 // +-------------------------------------------------+
-// $Id: editor.class.php,v 1.45.2.2 2014-07-31 09:11:22 dgoron Exp $
+// $Id: editor.class.php,v 1.54 2015-06-05 13:04:00 dgoron Exp $
 
 if (stristr($_SERVER['REQUEST_URI'], ".class.php")) die("no access");
 
@@ -15,12 +15,14 @@ require_once("$class_path/aut_link.class.php");
 require_once("$class_path/aut_pperso.class.php");
 require_once("$class_path/audit.class.php");
 require_once($class_path."/synchro_rdf.class.php");
+require_once($class_path."/index_concept.class.php");
+require_once($class_path."/vedette/vedette_composee.class.php");
 
 class editeur {
 
-// ---------------------------------------------------------------
-//		proprietes de la classe
-// ---------------------------------------------------------------
+	// ---------------------------------------------------------------
+	//		proprietes de la classe
+	// ---------------------------------------------------------------
 
 	var $id;			// MySQL id in table 'publishers'
 	var	$name;			// publisher name
@@ -72,10 +74,10 @@ class editeur {
 			$this->ed_comment	= '';
 		} else {
 			$requete = "SELECT * FROM publishers WHERE ed_id=$this->id LIMIT 1 ";
-			$result = @mysql_query($requete, $dbh);
-			if(mysql_num_rows($result)) {
-				$temp = mysql_fetch_object($result);
-				mysql_free_result($result);
+			$result = @pmb_mysql_query($requete, $dbh);
+			if(pmb_mysql_num_rows($result)) {
+				$temp = pmb_mysql_fetch_object($result);
+				pmb_mysql_free_result($result);
 				$this->id		= $temp->ed_id;
 				$this->name		= $temp->ed_name;
 				$this->adr1		= $temp->ed_adr1;
@@ -144,6 +146,7 @@ class editeur {
 		global $publisher_form;
 	 	global $charset;
 		global $pmb_type_audit;
+		global $thesaurus_concepts_active;
 	
 		if($this->id) {
 			$action = "./autorites.php?categ=editeurs&sub=update&id=$this->id";
@@ -188,8 +191,13 @@ class editeur {
 		$publisher_form = str_replace('!!user_input!!',			htmlentities($user_input,ENT_QUOTES, $charset),		$publisher_form);
 		$publisher_form = str_replace('!!nbr_lignes!!',			$nbr_lignes,										$publisher_form);
 		$publisher_form = str_replace('!!page!!',				$page,												$publisher_form);
-		$publisher_form = str_replace('!!ed_comment!!',		$this->ed_comment,									$publisher_form);
-		
+		$publisher_form = str_replace('!!ed_comment!!',			$this->ed_comment,									$publisher_form);
+		if($thesaurus_concepts_active == 1){
+			$index_concept = new index_concept($this->id, TYPE_PUBLISHER);
+			$publisher_form = str_replace('!!concept_form!!',		$index_concept->get_form('saisie_editeur'),		$publisher_form);
+		}else{
+			$publisher_form = str_replace('!!concept_form!!',		"",			$publisher_form);
+		}
 		if ($pmb_type_audit && $this->id)
 			$bouton_audit= "&nbsp;<input class='bouton' type='button' onClick=\"openPopUp('./audit.php?type_obj=".AUDIT_PUBLISHER."&object_id=".$this->id."', 'audit_popup', 700, 500, -2, -2, 'scrollbars=yes, toolbar=no, dependent=yes, resizable=yes')\" title=\"".$msg['audit_button']."\" value=\"".$msg['audit_button']."\" />&nbsp;";	
 		$publisher_form = str_replace('!!audit_bt!!',$bouton_audit,	$publisher_form);
@@ -232,24 +240,37 @@ class editeur {
 		$requete = "SELECT COUNT(1) FROM notices WHERE ";
 		$requete .= "ed1_id=$this->id OR ";
 		$requete .= "ed2_id=$this->id";
-		$res = mysql_query($requete, $dbh);
-		$nbr_lignes = mysql_result($res, 0, 0);
+		$res = pmb_mysql_query($requete, $dbh);
+		$nbr_lignes = pmb_mysql_result($res, 0, 0);
 		if(!$nbr_lignes) {
 			// on regarde si l'editeur a des collections enfants 
 			$requete = "SELECT COUNT(1) FROM collections WHERE ";
 			$requete .= "collection_parent=".$this->id;
-			$res = mysql_query($requete, $dbh);
-			$nbr_lignes = mysql_result($res, 0, 0);
+			$res = pmb_mysql_query($requete, $dbh);
+			$nbr_lignes = pmb_mysql_result($res, 0, 0);
 			if(!$nbr_lignes) {
+				
+				// On regarde si l'autorité est utilisée dans des vedettes composées
+				$attached_vedettes = vedette_composee::get_vedettes_built_with_element($this->id, "publisher");
+				if (count($attached_vedettes)) {
+					// Cette autorité est utilisée dans des vedettes composées, impossible de la supprimer
+					return '<strong>'.$this->name."</strong><br />".$msg["vedette_dont_del_autority"];
+				}
+				
 				// effacement dans la table des editeurs
 				$requete = "DELETE FROM publishers WHERE ed_id=".$this->id;
-				$result = mysql_query($requete, $dbh);
+				$result = pmb_mysql_query($requete, $dbh);
 				// liens entre autorités
 				$aut_link= new aut_link(AUT_TABLE_PUBLISHERS,$this->id);
 				$aut_link->delete();
 				
 				$aut_pperso= new aut_pperso("publisher",$this->id);
 				$aut_pperso->delete();
+				
+				// nettoyage indexation concepts
+				$index_concept = new index_concept($this->id, TYPE_PUBLISHER);
+				$index_concept->delete();
+				
 				audit::delete_audit(AUDIT_PUBLISHER,$this->id);
 				return false;
 			} else {
@@ -291,17 +312,17 @@ class editeur {
 		
 		// a) remplacement dans les notices
 		$requete = "UPDATE notices SET ed1_id=$by WHERE ed1_id=".$this->id;
-		$res = mysql_query($requete, $dbh);
+		$res = pmb_mysql_query($requete, $dbh);
 		$requete = "UPDATE notices SET ed2_id=$by WHERE ed2_id=".$this->id;
-		$res = mysql_query($requete, $dbh);
+		$res = pmb_mysql_query($requete, $dbh);
 	
 		// b) remplacement dans la table des collections
 		$requete = "UPDATE collections SET collection_parent=$by WHERE collection_parent=".$this->id;
-		$res = mysql_query($requete, $dbh);
+		$res = pmb_mysql_query($requete, $dbh);
 	
 		// c) suppression de l'editeur a remplacer
 		$requete = "DELETE FROM publishers WHERE ed_id=".$this->id;
-		$res = mysql_query($requete, $dbh);
+		$res = pmb_mysql_query($requete, $dbh);
 		
 		audit::delete_audit (AUDIT_PUBLISHER, $this->id) ;
 	
@@ -325,6 +346,7 @@ class editeur {
 		global $msg;
 		global $include_path;
 		global $pmb_synchro_rdf;
+		global $thesaurus_concepts_active;
 		
 		if(!$value['name'])
 			return false;
@@ -352,7 +374,7 @@ class editeur {
 			// update
 			$requete = 'UPDATE publishers '.$requete;
 			$requete .= ' WHERE ed_id='.$this->id.' LIMIT 1;';
-			if(mysql_query($requete, $dbh)) {
+			if(pmb_mysql_query($requete, $dbh)) {
 				$aut_link= new aut_link(AUT_TABLE_PUBLISHERS,$this->id);
 				$aut_link->save_form();
 				$aut_pperso= new aut_pperso("publisher",$this->id);
@@ -366,8 +388,6 @@ class editeur {
 					$synchro_rdf = new synchro_rdf();
 					$synchro_rdf->updateAuthority($this->id,'editeur');
 				}
-				
-				return TRUE;
 			}else {
 				require_once("$include_path/user_error.inc.php");
 				warning($msg[145], $msg[150]);
@@ -382,20 +402,27 @@ class editeur {
 				return FALSE;
 			}
 			$requete = 'INSERT INTO publishers '.$requete.';';
-			if(mysql_query($requete, $dbh)) {
-				$this->id=mysql_insert_id();
+			if(pmb_mysql_query($requete, $dbh)) {
+				$this->id=pmb_mysql_insert_id();
 				$aut_link= new aut_link(AUT_TABLE_PUBLISHERS,$this->id);
 				$aut_link->save_form();
 				$aut_pperso= new aut_pperso("publisher",$this->id);
 				$aut_pperso->save_form();
 				audit::insert_creation (AUDIT_PUBLISHER, $this->id) ;
-				return TRUE;
 			} else {
 				require_once("$include_path/user_error.inc.php");
 				warning($msg[145], $msg[151]);
 				return FALSE;
 			}
 		}
+		if($thesaurus_concepts_active == 1){
+			$index_concept = new index_concept($this->id, TYPE_PUBLISHER);
+			$index_concept->save();
+		}
+
+		// Mise à jour des vedettes composées contenant cette autorité
+		vedette_composee::update_vedettes_built_with_element($this->id, "publisher");
+		return TRUE;
 	}
 	
 	// ---------------------------------------------------------------
@@ -412,7 +439,7 @@ class editeur {
 	
 		// tentative de recuperer l'id associee dans la base (implique que l'autorite existe)
 		// preparation de la requeªte
-		$long_maxi = mysql_field_len(mysql_query("SELECT ed_name FROM publishers limit 1"),0);
+		$long_maxi = pmb_mysql_field_len(pmb_mysql_query("SELECT ed_name FROM publishers limit 1"),0);
 		
 		$key = addslashes(rtrim(substr(preg_replace('/\[|\]/', '', rtrim(ltrim($data['name']))),0,$long_maxi)));
 		$ville=addslashes(trim($data['ville']));
@@ -426,12 +453,12 @@ class editeur {
 		if ($key=="") return 0; /* on laisse tomber les editeurs sans nom !!! exact. FL*/
 	
 		$query = "SELECT ed_id FROM publishers WHERE ed_name='${key}' and ed_ville = '${ville}' ";
-		$result = @mysql_query($query, $dbh);
+		$result = @pmb_mysql_query($query, $dbh);
 		if(!$result) die("can't SELECT publisher ".$query);
 		// resultat
 	
 		// recuperation du resultat de la recherche
-		$tediteur  = mysql_fetch_object($result);
+		$tediteur  = pmb_mysql_fetch_object($result);
 		// et recuperation eventuelle de l'id
 		if($tediteur->ed_id)
 			return $tediteur->ed_id;
@@ -440,9 +467,9 @@ class editeur {
 	
 		$query = "INSERT INTO publishers SET ed_name='${key}', ed_ville = '${ville}', ed_adr1 = '${adr}', ed_comment='".$ed_comment."', ed_adr2='".$adr2."', ed_cp='".$cp."', ed_pays='".$pays."', ed_web='".$web."', index_publisher=' ".strip_empty_chars($key)." ' ";
 	
-		$result = @mysql_query($query, $dbh);
+		$result = @pmb_mysql_query($query, $dbh);
 		if(!$result) die("can't INSERT into publisher : ".$query);
-		$id=mysql_insert_id($dbh);
+		$id=pmb_mysql_insert_id($dbh);
 		
 		audit::insert_creation (AUDIT_PUBLISHER, $id) ;
 		return $id;
@@ -470,12 +497,12 @@ class editeur {
 	//---------------------------------------------------------------
 	// update_index($id) : maj des n-uplets la table notice_global_index en rapport avec cet editeur	
 	//---------------------------------------------------------------
-	function update_index($id) {
+	static function update_index($id) {
 		global $dbh;
 		// On cherche tous les n-uplet de la table notice correspondant a cet auteur.
-		$found = mysql_query("select distinct notice_id from notices where ed1_id='".$id."' OR ed2_id='".$id."'",$dbh);
+		$found = pmb_mysql_query("select distinct notice_id from notices where ed1_id='".$id."' OR ed2_id='".$id."'",$dbh);
 		// Pour chaque n-uplet trouves on met a jour la table notice_global_index avec l'auteur modifie :
-		while($mesNotices = mysql_fetch_object($found)) {
+		while($mesNotices = pmb_mysql_fetch_object($found)) {
 			$notice_id = $mesNotices->notice_id;
 			notice::majNoticesGlobalIndex($notice_id);
 			notice::majNoticesMotsGlobalIndex($notice_id,'publisher');
@@ -485,7 +512,7 @@ class editeur {
 	//---------------------------------------------------------------
 	// get_informations_from_unimarc : ressort les infos d'un éditeur depuis une notice unimarc
 	//---------------------------------------------------------------
-	function get_informations_from_unimarc($fields){
+	static function get_informations_from_unimarc($fields){
 		$data = array();
 		if($fields['210']){
 			$data['name'] = $fields['210'][0]['c'][0];
@@ -499,19 +526,19 @@ class editeur {
 	static function check_if_exists($data){
 		global $dbh;
 		
-		$long_maxi = mysql_field_len(mysql_query("SELECT ed_name FROM publishers limit 1"),0);
+		$long_maxi = pmb_mysql_field_len(pmb_mysql_query("SELECT ed_name FROM publishers limit 1"),0);
 		$key = addslashes(rtrim(substr(preg_replace('/\[|\]/', '', rtrim(ltrim($data['name']))),0,$long_maxi)));
 		$ville=addslashes(trim($data['ville']));
 		$adr=addslashes(trim($data['adr']));
 		$ed_comment=addslashes(trim($data['ed_comment']));
 		
 		$query = "SELECT ed_id FROM publishers WHERE ed_name='${key}' and ed_ville = '${ville}' ";
-		$result = @mysql_query($query, $dbh);
+		$result = @pmb_mysql_query($query, $dbh);
 		if(!$result) die("can't SELECT publisher ".$query);
 		// resultat
 	
 		// recuperation du resultat de la recherche
-		$tediteur  = mysql_fetch_object($result);
+		$tediteur  = pmb_mysql_fetch_object($result);
 		// et recuperation eventuelle de l'id
 		if($tediteur->ed_id)
 			return $tediteur->ed_id;
